+++ pop_pay/__init__.py
@@ -1,6 +1,6 @@
 from importlib.metadata import version, PackageNotFoundError
 
 try:
-    __version__ = version("aegis-pay")
+    __version__ = version("pop-pay")
 except PackageNotFoundError:
     __version__ = "unknown"